fix(Recette): Correction d'un bug ne permettant pas de sauvegarder les ingredients et les etapes lors du save d'une recette

+++ api/src/main/java/com/api/kookie/core/profile/ProfileService.java
@@ -12,7 +12,7 @@ public interface ProfileService {
 
     ProfileDTO getByProfileId(Integer profileId);
 
-    Boolean addFavorite(Integer profileId, Long recetteId);
+    Boolean addFavorite(Integer profileId, Integer recetteId);
 
     List<RecetteDTO> getFavorites(Integer profileId);
 }
